requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
pillow==10.3.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
idna==3.7
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
drf-spectacular==0.27.2
django-filter==24.2
django-jazzmin==3.0.0
django-cloudinary-storage==0.3.0
cloudinary==1.40.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
celery==5.2.7
celery[redis]
django-redis==5.4.0
redis==5.0.1

